% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_color_orders.R
\name{custom_discrete_viridis_palette}
\alias{custom_discrete_viridis_palette}
\title{Discretized Luminance Linear Color Gradient}
\usage{
custom_discrete_viridis_palette(n)
}
\description{
Discretized Luminance Linear Color Gradient
}
