% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{edit_the_accent_palette}
\alias{edit_the_accent_palette}
\title{Modify supplemental colors in theme}
\usage{
edit_the_accent_palette(..., mode = c("overwrite", "append"))
}
\arguments{
\item{...}{A vector or comma-separated list of colors}

\item{mode}{\code{"overwrite"} (the default), removes all existing accent colors
and replace them with \code{...}. The \code{"append"} option leaves all existing
colors and adds every color specified in \code{...}.}
}
\description{
The accent palette contains an arbitrary number of additional colors for your
theme that can be used when coloring discrete/categorical variables. The
accent palette does not have meaningful names or an order. Because of this,
by default, \code{edit_the_accent_palette} overwrites \emph{the entire} existing accent
palette. You can append to the existing list by using \code{mode = 'append'}
argument. Behind the scenes, ggchameleon figures out which colors are most
differentiable, and includes those in your chart.
}
\examples{
# Add additional colors to the accent palette
edit_the_accent_palette("tomato","powderblue",mode="append")

# Overwrite the existing accent colors and replace with new colors
edit_the_accent_palette(c("orchid","lavender","violet","goldenrod"))

}
