% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{edit_the_fonts}
\alias{edit_the_fonts}
\title{Use custom fonts from Google Fonts}
\usage{
edit_the_fonts(...)
}
\arguments{
\item{...}{Any fonts available in \code{the$fonts}}
}
\description{
\code{edit_the_fonts} allows you specify any fonts available in the \href{https://fonts.google.com/}{Google Fonts} library as fonts for your ggchameleon
configurations. You can specify different fonts depending on which part of
the chart you are interested in modifying. \code{edit_the_fonts} defers to
\code{\link[sysfonts:font_add_google]{sysfonts::font_add_google()}} for finding and installing Google fonts.
}
\examples{
# Changes the title font to the "Lobster" font family
edit_the_fonts(display="Lobster")

}
