% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{edit_the_main_palette}
\alias{edit_the_main_palette}
\title{Modify the primary colors used in your charts}
\usage{
edit_the_main_palette(...)
}
\arguments{
\item{...}{Any color that is available in \code{the$main_palette}}
}
\description{
\code{edit_the_main_palette()} allows you to change the colors that automatically
appear in your ggplots. When you edit a color, ggchameleon will figure out
where those new colors should appear in your charts. ggchameleon will not
only update the text colors, but also the default colors of the geometries
(the shapes that appear in the chart) when applicable. You can edit one or
more parameters at a time. After editing parameters, the updated values are
stored in \link{the}.
}
\examples{
# Editing one parameter at a time
# converts the main color to magenta (pink)
edit_the_main_palette(main="magenta")
# Editing multiple elements
# continuous gradients now go from green to gray to magenta
edit_the_main_palette(contrast="darkolivegreen",intermediate="gray50")

# our changes are also reflected in `the`
the$main_palette

}
