% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{edit_the_theme}
\alias{edit_the_theme}
\title{Specify a custom ggplot2 plot layout}
\usage{
edit_the_theme(...)
}
\arguments{
\item{...}{Arguments to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\description{
With \code{edit_the_theme}, anything that can be modified with \code{\link[ggplot2:theme]{ggplot2::theme()}}
syntax can be set as a default under the ggchameleon framework. The syntax of
\code{edit_the_theme} mirrors that of \code{\link[ggplot2:theme]{ggplot2::theme()}}. You can customize layout
parameters by customizing options like \code{legend.position}, or you can
customize options related to aesthetics, like the font or colors.
Importantly, \code{edit_the_theme} is unable to customize the "geometries" (the
shapes representing the data), this is purely the supplemental shapes of a
ggplot.
}
\examples{
# Places the legend at the bottom by default.
edit_the_theme(legend.position='bottom')

# Change the title text color to the secondary color
edit_the_theme(title = element_text(color = the$main_palette$secondary))

}
