% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_check.R
\name{font_check}
\alias{font_check}
\title{Check if all fonts are installed and install if not}
\usage{
font_check()
}
\value{
nothing, but messages if no fonts are installed
}
\description{
Look for fonts in relevant directory. If none are found, then install those fonts.
}
\details{
You cannot directly edit any other theme elements. This function returns a complete theme, and if you wish to further modify you should use the ggplot2::theme() command
}
\examples{
font_check()



}
