% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_color_orders.R
\name{gen_palette}
\alias{gen_palette}
\title{Generate Maximally Differentiable Set of Palette Colors:}
\usage{
gen_palette(n)
}
\description{
Generate Maximally Differentiable Set of Palette Colors:
}
