% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huemint_api.R
\name{huemint_randomize}
\alias{huemint_randomize}
\title{Randomize Colors to a (Usually) Pleasant Palette}
\usage{
huemint_randomize(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean. Whether or not to preview the new palette. Defaults
to \code{TRUE}.}
}
\description{
\code{huemint_randomize} \emph{overwrites} the existing colors in \link{the} (the
environment that stores your current configurations). \code{huemint_randomize}
leverages \link{https://huemint.com/} to generate the new color palette. Huemint
uses machine learning to create color palettes that are novel and appealing.
You can read more about that process \href{https://huemint.com/about/}{here}.

\code{huemint_randomize} imposes further restrictions on the generated palette---
requiring certain colors to be more and less differentiable. For example the
generated black should look very different from the generated white.
}
