% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_custom_fonts.R
\name{install_fonts}
\alias{install_fonts}
\title{Get fonts that resemble the fonts we use for publication}
\usage{
install_fonts(fonts = the$fonts)
}
\value{
Nothing,
}
\description{
Pulls fonts from google then installs them into a file in the package location and then adds them to the list of R-available fonts
}
\details{
You cannot directly edit any other theme elements. This function returns a complete theme, and if you wish to further modify you should use the ggplot2::theme() command
}
\examples{
install_philly_fonts()

}
