% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_configs.R
\name{save_configs}
\alias{save_configs}
\title{Save stylistic configurations to a file for later use}
\usage{
save_configs(file = "chameleon.yml", overwrite = FALSE)
}
\arguments{
\item{file}{A file path ending in \code{".yml"}. Defaults to \code{"chameleon.yml"}.}

\item{overwrite}{A boolean. Whether or not to overwrite \code{file} if it exists.
Defaults to \code{FALSE}. When \code{FALSE}, \code{file} will have random characters
appended to the end of the name to disambiguate.}
}
\description{
If you are happy with your current chart style, and you wish to use it again
in the future, you can save the current state of the ggchameleon
configurations to a file on your machine using \code{save_configs()}. By default,
this creates a file called \code{chameleon.yml} which stores the fonts, colors,
and \code{\link[ggplot2:theme]{ggplot2::theme()}} modifications you have made. After you end your R
session, you can reload these stylings using the \code{\link[=load_configs]{load_configs()}} function.
}
\details{
By default, \code{save_configs()} creates a file named \code{chameleon.yml} in your current working directory.
If you are using an R project (ie if you have a \code{.Rproj} file), then your working directory is
likely wherever the \code{.Rproj} located. When ggchameleon is loaded, it automatically looks for
this \code{chameleon.yml} file, and if it finds it, it will automatically apply those
configurations. This is convenient because whenever you re-run any code with
ggplot2, it will automatically start using whatever you have saved.
}
\examples{

# Save the configs file to the default location
save_configs()

# save the configs to a different file name
save_configs("my_different_configs.yml")

}
