% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colors.R
\name{show_colors}
\alias{show_colors}
\title{Show the philly color pallete, or basically any color pallette
Stolen from scales::show_col with the modification to allow for named color lists}
\usage{
show_colors(
  colours = c(the$main_palette, the$accent_palette),
  labels = TRUE,
  borders = NULL,
  cex_label = 1,
  ncol = NULL
)
}
\description{
Show the philly color pallete, or basically any color pallette
Stolen from scales::show_col with the modification to allow for named color lists
}
