% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_custom.R
\name{theme_custom}
\alias{theme_custom}
\title{Add Custom Theme to a ggplot}
\usage{
theme_custom(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family, defaults to COMPLETE THIS}

\item{base_line_size}{bae size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
A ggplot2 theme object used to update a ggplot2 plot.
}
\description{
This function is used exactly like a regular ggplot2 theme object. It can be added to a ggplot to change the stylistic elements of a plot
}
\details{
You cannot directly edit any other theme elements. This function returns a complete theme, and if you wish to further modify you should use the \code{\link[ggplot2:theme]{ggplot2::theme()}} command
}
