% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{edit_the_main_palette}
\alias{edit_the_main_palette}
\title{Modify the primary colors used in your charts}
\usage{
edit_the_main_palette(...)
}
\arguments{
\item{main}{The primary color used for title text and appearing whenever
possible in charts}

\item{secondary}{A color that should notably contrast the \code{main} color}

\item{white}{The whitest color that will appear in your charts}

\item{off_white}{The color used for grid lines, subtly differentiable from
\code{white}}

\item{black}{The darkest color that will appear in your plots}

\item{contrast}{A color that is \strong{very} differentiable from \code{main}. Used
whenever there is a continuous color gradient}

\item{intermediate}{A color to appear between \code{main} and \code{contrast} in
continuous gradients}
}
\description{
\code{edit_the_main_palette()} allows you to change the colors that automatically
appear in your ggplots. When you edit a color, ggchameleon will figure out
where those new colors should appear in your charts. ggchameleon will not
only update the text colors, but also the default colors of the geometries
(the shapes that appear in the chart) when applicable. You can edit one or
more parameters at a time. After editing parameters, the updated values are
stored in \link{the}.
}
\examples{
# Editing one parameter at a time
# converts the main color to magenta (pink)
edit_the_main_palette(main="magenta")
# Editing multiple elements
# continuous gradients now go from green to gray to magenta
edit_the_main_palette(contrast="darkolivegreen",intermediate="gray50")

# our changes are also reflected in `the`
the$main_palette

}
